chore: remove shims, add tests, Makefile targets, and architecture doc
