+++ app/static/js/ui/windows.js
@@ -3,6 +3,5 @@ export const windows = [
   { id: "win_docs",     window_type: "window_documents", title: "Document Library", col: "left"  },
   { id: "win_sessions", window_type: "window_sessions",  title: "Chat History",     col: "left"  },
   { id: "win_chat",     window_type: "window_chat_ui",   title: "Assistant Chat",   col: "right" },
-  { id: "win_search",   window_type: "window_search",    title: "Search Documents", col: "right" },
   { id: "win_segments", window_type: "window_segments",  title: "DB Segments",      col: "right" }
 ];
