+++ app/templates/index.html
@@ -16,6 +16,7 @@
         <div id="menu-dropdown" class="menu-dropdown" role="menu" aria-hidden="true">
           <button class="menu-item" data-action="new-chat" role="menuitem">New Chat</button>
           <button class="menu-item" data-action="edit-persona" role="menuitem">Persona…</button>
+          <button class="menu-item" data-action="toggle-search" role="menuitem">Context</button>
           <button class="menu-item" data-action="settings" role="menuitem">Settings…</button>
         </div>
       </div>
