+++ core/rag/embeddings.py
@@ -1,4 +1,4 @@
-# utility/model_utils.py
+from __future__ import annotations
 from functools import lru_cache
 from pathlib import Path
 import os
@@ -6,20 +6,19 @@ from typing import Optional
 from sentence_transformers import SentenceTransformer
 
 try:
-    # Optional: only needed for the one-time fetch command
-    from huggingface_hub import snapshot_download
+    from huggingface_hub import snapshot_download  # optional
 except Exception:
     snapshot_download = None
 
 from config import MODEL_DIR
 
-# New: specify both ID and on-disk dir
 EMBEDDING_MODEL_ID = os.getenv("EMBEDDING_MODEL_ID", "sentence-transformers/all-MiniLM-L6-v2")
 EMBEDDINGS_OFFLINE_ONLY = os.getenv("EMBEDDINGS_OFFLINE_ONLY", "0") == "1"
 
+
 def _local_model_present(model_dir: Path) -> bool:
-    # crude but effective: directory exists and has files
     return model_dir.exists() and any(model_dir.iterdir())
+
 
 def fetch_model_if_needed(model_id: str = EMBEDDING_MODEL_ID, model_dir: Path = MODEL_DIR) -> Path:
     if _local_model_present(model_dir):
@@ -32,10 +31,11 @@ def fetch_model_if_needed(model_id: str = EMBEDDING_MODEL_ID, model_dir: Path = 
         repo_id=model_id,
         local_dir=str(model_dir),
         local_dir_use_symlinks=False,
-        allow_patterns=None,  # full repo
-        ignore_patterns=["*.safetensors.index.json"],  # optional
+        allow_patterns=None,
+        ignore_patterns=["*.safetensors.index.json"],
     )
     return model_dir
+
 
 @lru_cache(maxsize=1)
 def load_embedding_model(force_fetch: bool = False) -> SentenceTransformer:
@@ -46,5 +46,4 @@ def load_embedding_model(force_fetch: bool = False) -> SentenceTransformer:
         if EMBEDDINGS_OFFLINE_ONLY:
             raise RuntimeError(f"Model cache missing at {model_dir} and offline-only is enabled.")
         fetch_model_if_needed()
-    # Always load from local disk
     return SentenceTransformer(str(model_dir), device=os.getenv("EMBEDDINGS_DEVICE", "cpu"))
