+++ core/sessions.py
@@ -1,35 +1,21 @@
-"""Chat session and exchange data models.
-
-This module defines lightweight data containers used to persist chat
-history and session level metadata.  The previous implementation used
-ad-hoc classes with a mixture of attribute names (e.g. ``llm_response``)
-which made it difficult to work with the objects consistently.  The
-refactored version below leverages :mod:`dataclasses` to provide a
-clear, typed structure and introduces a single ``assistant`` field for
-LLM responses.  For backwards compatibility the loader understands the
-old ``llm_response`` key.
-"""
-
+"""Chat session models and persistence."""
 from __future__ import annotations
-
-import uuid
 from dataclasses import dataclass, field
 from typing import List, Optional, Dict
+import uuid
+import json
+from pathlib import Path
 
+# --- Chat history models ---
 
 @dataclass
 class ChatExchange:
-    """Represents a single interaction within a chat session."""
-
     user: str
     context_used: List[Dict]
     rag_prompt: str
     assistant: str
     html_response: str
 
-    # ------------------------------------------------------------------
-    # Serialization helpers
-    # ------------------------------------------------------------------
     def to_dict(self) -> Dict:
         return {
             "user": self.user,
@@ -41,12 +27,6 @@ class ChatExchange:
 
     @classmethod
     def from_dict(cls, data: Dict) -> "ChatExchange":
-        """Create an exchange from serialized data.
-
-        Both ``assistant`` and the legacy ``llm_response`` keys are
-        supported when restoring older session files.
-        """
-
         return cls(
             user=data.get("user", ""),
             context_used=data.get("context_used", []),
@@ -55,11 +35,8 @@ class ChatExchange:
             html_response=data.get("html_response", ""),
         )
 
-
 @dataclass
 class ChatSession:
-    """Container for a user's ongoing chat session."""
-
     session_id: str
     user_id: str = "default"
     history: List[ChatExchange] = field(default_factory=list)
@@ -68,54 +45,24 @@ class ChatSession:
     inactive_sources: List[str] = field(default_factory=list)
     persona: Optional[str] = None
 
-    # ------------------------------------------------------------------
-    # Construction helpers
-    # ------------------------------------------------------------------
     @classmethod
     def new(cls, session_id: Optional[str] = None, user_id: str = "default") -> "ChatSession":
-        """Create a new chat session with a random identifier."""
-
         return cls(session_id=session_id or str(uuid.uuid4()), user_id=user_id)
 
-    # ------------------------------------------------------------------
-    # Mutation helpers
-    # ------------------------------------------------------------------
-    def add_exchange(
-        self,
-        user: str,
-        context_used: List[Dict],
-        rag_prompt: str,
-        assistant: str,
-        html_response: str,
-    ) -> None:
-        """Append a new exchange to the session history."""
-
-        self.history.append(
-            ChatExchange(
-                user=user,
-                context_used=context_used,
-                rag_prompt=rag_prompt,
-                assistant=assistant,
-                html_response=html_response,
-            )
-        )
+    def add_exchange(self, user: str, context_used: List[Dict], rag_prompt: str, assistant: str, html_response: str) -> None:
+        self.history.append(ChatExchange(user, context_used, rag_prompt, assistant, html_response))
 
     def trim_history(self, max_length: int) -> None:
-        """Keep only the most recent ``max_length`` exchanges."""
-
         if len(self.history) > max_length:
             self.history = self.history[-max_length:]
 
-    # ------------------------------------------------------------------
-    # Serialization helpers
-    # ------------------------------------------------------------------
     def to_dict(self) -> Dict:
         return {
             "session_id": self.session_id,
             "user_id": self.user_id,
             "summary": self.summary,
             "title": self.title,
-            "history": [exchange.to_dict() for exchange in self.history],
+            "history": [h.to_dict() for h in self.history],
             "inactive_sources": self.inactive_sources,
             "persona": self.persona,
         }
@@ -133,3 +80,56 @@ class ChatSession:
         session.history = [ChatExchange.from_dict(e) for e in data.get("history", [])]
         return session
 
+# --- Session store ---
+
+SESSION_DIR = Path("sessions")
+SESSION_DIR.mkdir(exist_ok=True)
+
+class SessionStore:
+    def __init__(self, storage_path: Path = SESSION_DIR):
+        self.storage_path = storage_path
+
+    def _session_path(self, session_id: str) -> Path:
+        return self.storage_path / f"{session_id}.json"
+
+    def save(self, session: ChatSession) -> None:
+        with open(self._session_path(session.session_id), "w", encoding="utf-8") as f:
+            json.dump(session.to_dict(), f, indent=2)
+
+    def load(self, session_id: str) -> Optional[ChatSession]:
+        path = self._session_path(session_id)
+        if not path.exists():
+            return None
+        try:
+            with open(path, "r", encoding="utf-8") as f:
+                data = json.load(f)
+            return ChatSession.from_dict(data)
+        except json.JSONDecodeError:
+            path.unlink()
+            return None
+
+    def list_sessions(self) -> List[Dict]:
+        entries = []
+        for f in self.storage_path.glob("*.json"):
+            entries.append(
+                {
+                    "id": f.stem,
+                    "path": str(f),
+                    "modified": f.stat().st_mtime,
+                }
+            )
+        return entries
+
+    def delete(self, session_id: str) -> None:
+        p = self._session_path(session_id)
+        if p.exists():
+            p.unlink()
+
+    def exists(self, session_id: str) -> bool:
+        return self._session_path(session_id).exists()
+
+    def prune_empty(self) -> None:
+        for entry in list(self.list_sessions()):
+            session = self.load(entry["id"])
+            if session and not session.history:
+                self.delete(entry["id"])
