+++ makefile
@@ -14,7 +14,7 @@ APP_MODULE?= app.main:app
 MODEL_ID  ?= sentence-transformers/all-MiniLM-L6-v2
 
 
-.PHONY: setup venv install fetch-model verify-offline run embed-dir clean
+.PHONY: setup venv install fetch-model verify-offline run dev embed-dir clean test seed-prompts
 
 # ---------- ONLINE SETUP ----------
 setup: export TRANSFORMERS_OFFLINE=0
@@ -32,31 +32,40 @@ install:
 # Cache the embedding model into config.MODEL_DIR
 fetch-model:
 	@echo "📥 Caching embedding model ($(MODEL_ID))..."
-	@PYTHONPATH=. $(PY) -c "from utility.model_utils import load_embedding_model; load_embedding_model(True); print('✓ cached via model_utils')" || \
+	@PYTHONPATH=. $(PY) -c "from core.rag.embeddings import load_embedding_model; load_embedding_model(True); print('✓ cached via embeddings')" || \
 	( echo 'model_utils failed; falling back to huggingface_hub…' && \
 	  PYTHONPATH=. $(PYTHON) -c "from huggingface_hub import snapshot_download; from config import MODEL_DIR as MD; snapshot_download(repo_id='$(MODEL_ID)', local_dir=str(MD), local_dir_use_symlinks=False); print('✓ cached under', MD)" )
 
 # ---------- OFFLINE CHECK ----------
 verify-offline:
 	@if [ -x "$(PY)" ]; then PYBIN="$(PY)"; else echo "⚠️  $(PY) missing; falling back to $(PYTHON)"; PYBIN="$(PYTHON)"; fi; \
-	PYTHONPATH=. $$PYBIN -c "from utility.model_utils import load_embedding_model as L; m=L(); print('offline OK:', m.get_sentence_embedding_dimension())" || \
+	PYTHONPATH=. $$PYBIN -c "from core.rag.embeddings import load_embedding_model as L; m=L(); print('offline OK:', m.get_sentence_embedding_dimension())" || \
 	PYTHONPATH=. $$PYBIN -c "from sentence_transformers import SentenceTransformer; from config import MODEL_DIR as MD; SentenceTransformer(str(MD)); print('offline OK via direct local model path ✓')"
 
 # ---------- RUN ----------
 run: verify-offline
 	@if [ -x "$(UVICORN)" ]; then \
-	  $(UVICORN) $(APP_MODULE) --host 127.0.0.1 --port 8000 --reload; \
+	  $(UVICORN) $(APP_MODULE) --host 127.0.0.1 --port 8000; \
 	else \
 	  echo "❌ $(UVICORN) not found. Run 'make setup' (online) first."; \
 	  exit 127; \
 	fi
 
+dev:
+	$(UVICORN) $(APP_MODULE) --host 127.0.0.1 --port 8000 --reload
+
 # ---------- OPTIONAL: batch embed without API ----------
 embed-dir:
 	@echo "Embedding from documents/ (override with: make embed-dir DATA_DIR=path)"
-	@PYTHONPATH=. $(PY) embedding_and_storing.py --data_dir "$${DATA_DIR:-documents}"
+	@PYTHONPATH=. $(PY) -c "from core.rag.retriever import embed_directory; embed_directory(data_dir='$${DATA_DIR:-documents}')"
 
 # ---------- housekeeping ----------
 clean:
 	rm -rf $(VENV_NAME)
 	rm -rf chroma/
+
+test:
+	pytest -q
+
+seed-prompts:
+	@$(PYTHON) -c "from core.settings import list_prompt_templates as L; t=L(); assert t, 'no prompt templates found'; print('templates:', [x.id for x in t])"
